Move LogDivider to its own file to fix react-refresh lint warnings
